Move StringHeapData into heap/string.rs
